converted all files to ts, updated errors, except for __audits in test/basic.ts
