--- index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  parsers: require('./parsers'),
-  validators: require('./validators'),
-  auditer: require('./audit')
-}
