+++ index.ts
@@ -1,0 +1,10 @@
+import Parsers from "./parsers";
+import Auditers from "./audit";
+import Validators from "./validators";
+
+export default {
+  auditer: Auditers,
+  parsers: Parsers,
+  validators: Validators,
+}
+
